feature: add tests

+++ migrations/sqls/20241028134748-products-table-up.sql
@@ -1,1 +1,1 @@
-CREATE TABLE products (id SERIAL PRIMARY KEY, name VARCHAR(100), price MONEY, category VARCHAR(100));
+CREATE TABLE products (id SERIAL PRIMARY KEY, name VARCHAR(100), price INTEGER, category VARCHAR(100));
